refactor: STD-5 금액 관련 타입 변경

+++ src/main/java/com/tenten/studybadge/payment/dto/PaymentHistory.java
@@ -16,7 +16,7 @@ public class PaymentHistory {
 
     private LocalDateTime createdAt;
 
-    private Long amount;
+    private Integer amount;
 
     private String paymentKey;
